# Conflicts:
#	.vs/METAL/v17/.suo
#	.vs/METAL/v17/DocumentLayout.backup.json
#	.vs/METAL/v17/DocumentLayout.json
#	Assembly-CSharp.csproj
#	UserSettings/EditorUserSettings.asset

+++ Assets/Script/Enemy/Enemy_Shooter.cs
@@ -140,7 +140,7 @@ public class Enemy_Shooter : MonoBehaviour, Enemy_Manager
         ScoreManager.Instance.AddScore(scoreValue);
     }
 
-    void Enemy_Manager.TakeDamage(float damage)
+    void Enemy_Manager.TakeDamage(int damage)
     {
         TakeDamage(damage);
     }
